add securities

+++ scripts/xaal_activity_IR_cams.py
@@ -18,33 +18,35 @@ base_directory = "/run/user/1000/gvfs/smb-share:server=10.77.3.109,share=e/datas
 
 def start_activity(_activity):
 
-    global processus_ir_cam_big
-    global processus_ir_cam_small
+    if dev.attributes[0].value == True and dev.attributes[1].value != _activity:
 
-    if dev.attributes[0].value == True:
-        processus_ir_cam_big.terminate()
-        processus_ir_cam_small.terminate()
+        global processus_ir_cam_big
+        global processus_ir_cam_small
 
-    dev.attributes[0].value = True
-    dev.attributes[1].value = _activity
-    logger.debug(f"Starting recording {_activity}")
+        if dev.attributes[0].value == True:
+            processus_ir_cam_big.terminate()
+            processus_ir_cam_small.terminate()
 
-    path_big = base_directory+"/"+_activity+"/Optrix"
+        dev.attributes[0].value = True
+        dev.attributes[1].value = _activity
+        logger.debug(f"Starting recording {_activity}")
 
-    if not os.path.exists(path_big):
-        os.makedirs(path_big)
+        path_big = base_directory+"/"+_activity+"/Optrix"
 
-    processus_ir_cam_big = subprocess.Popen(args=["rosrun","optris_drivers","camera.py", "-a", _activity, "-d",path_big], stdout=subprocess.PIPE)
-    
+        if not os.path.exists(path_big):
+            os.makedirs(path_big)
 
-    path_small = base_directory+"/"+_activity+"/mini_IR"
-    filename = "/miniIR.mp4"
-    
-    if not os.path.exists(path_small):
-        os.makedirs(path_small)
+        processus_ir_cam_big = subprocess.Popen(args=["rosrun","optris_drivers","camera.py", "-a", _activity, "-d",path_big], stdout=subprocess.PIPE)
+        
 
-    processus_ir_cam_small = subprocess.Popen(args=["ffmpeg","-i","rtsp://xaal-c.enstb.org:8554/flir1", path_small+filename], stdout=subprocess.PIPE)
-    
+        path_small = base_directory+"/"+_activity+"/mini_IR"
+        filename = "/miniIR.mp4"
+        
+        if not os.path.exists(path_small):
+            os.makedirs(path_small)
+
+        processus_ir_cam_small = subprocess.Popen(args=["ffmpeg","-i","rtsp://xaal-c.enstb.org:8554/flir1", path_small+filename], stdout=subprocess.PIPE)
+        
 
 def stop_recording():
     logger.debug(f"Stop recording")
@@ -70,7 +72,8 @@ def main():
     dev.add_method('stop_recording',stop_recording)
     dev.new_attribute("state",None)
     dev.new_attribute("activity",None)
-    
+    dev.info = 'IR CAMS'
+
     eng = Engine()
     eng.add_device(dev)
     eng.run()
